requirements.txt
Flask==3.1.1
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.9.0
SQLAlchemy==2.0.42
StrEnum==0.4.15
Werkzeug==3.1.3
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
certifi==2025.7.14
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudevents==1.11.0
defusedxml==0.7.1
deprecation==2.1.0
et_xmlfile==2.0.0
fastapi==0.115.2
fonttools==4.59.0
fpdf2==2.8.3
frozenlist==1.7.0
functions-framework==3.9.2
gotrue==2.11.4
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
multidict==6.6.3
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
postgrest==0.16.11
propcache==0.3.2
psutil==6.1.1
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==21.4
pytz==2025.2
realtime==2.5.0
reportlab==4.2.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
starlette==0.40.0
storage3==0.7.7
supabase==2.7.4
supafunc==0.5.1
tenacity==9.0.0
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn-worker==0.3.0
uvicorn==0.32.0
watchdog==6.0.0
websockets==14.0
xlrd==2.0.2
yarl==1.20.1
matplotlib==3.9.4
